requirements.txt
moderage-python==0.2.2
pandas==0.24.2
matplotlib==3.1.1
numpy==1.18.1
pygame==1.9.6
cloudpickle==1.2.2
tensorboard==2.0.0
gym==0.15.6